Ensure Qt variables read on rebuild
When a Rebuild action is requested (i.e. Clean -> Build), the procedure
to extract Qt build settings by calling qmake will now be invoked at the
start of build. Previously, this procedure was not included in a
Rebuild, given that the generated files would not be deleted upon a
Clean request.

Change-Id: Ibfb28db44478f34f2b2024f5a2fbdb8c119f2948
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>