Enable lazy init of Qt version information
Information on registered Qt versions is loaded by means of querying
data from qmake. To avoid an inefficient initialization, the Qt version
information is loaded lazily, on an as-needed basis. This was already
implemented, by means of a cache in the QtVersionManager. However, it
had no effect, given that, at startup, the information of all Qt
versions is checked for validity, thereby greedily loading all version
infos. This is now fixed in the current patch: (1) when checking the
validity of Qt versions, only the presence of qmake.exe is actually
verified; (2) when selecting the latest version (as a default version,
if none is provided), will use the version information in qconfig.pri,
instead of querying qmake.

Change-Id: Ie270816250e775f7bdeb7cad9a290f07409fbc03
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>