Add support for cast of enum values
Enums in C# allow a compiler-verified way of maintaining lists of
possible values. Conversion of enums to other types is limited to the
underlying integer value of each enum value, and to the string
representation of the enum value name. For an alternative, customizable
conversion (e.g. to a string representation other than the value name),
it is possible to associate an attribute to an enum value, containing
the corresponding target type value/instance:

    enum Foobar {
        Foo,
        [EnumExt.String("Bahr")] Bar
    }

Converting an enum value will output the value in the associated
attribute, or a default conversion, if an attribute is not present.

    Foobar.Foo.Cast<string>()  -->  "Foo"
    Foobar.Bar.Cast<string>()  -->  "Bahr"

This extended enum conversion feature is used in upcoming patches.

Change-Id: Ifdd63c8bbb55f812772fdcbe146b7b01b16c57cc
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>