Fix deploy target not working for WinRT projects
The QtDeploy target will now provide correct information needed for
packaging a WinRT application. Specifically, a source root folder for
deploy items is now explicitly given. This defaults to the local path of
the MSBuild file that creates the item. If the deploy target is located
inside a project file (that was the case for WinRT projects created by
the VS add-in), the root folder path would correctly default to the
project directory. Using a deploy target located in Qt/MSBuild, the
default path is no longer valid and must be explicitly set to the
project directory.

Change-Id: Id8008d182184abfe0ee18ff61ff3e2abe1b0b686
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>