Add configurations wizard page
Added a wizard page to be used in new-project wizards for setting up the
build configurations available in newly created projects. This page will
allow users to create projects referencing multiple versions of Qt,
according to the build configuration. Previously, only a pair of
(Debug, Release) configurations would be added to new projects, with
reference to the default Qt version. The only customization that was
available was a selection of which Qt modules to use in the new project.

Task-number: QTVSADDINBUG-589
Task-number: QTVSADDINBUG-737
Change-Id: I72aa4039cbf0d81662ff66d10026a53064e32bc3
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>