Add base class for new-project wizards
Added a new class that contains features common to all new-project
templates. Specific wizards will inherit from this class, instead of
directly implementing the IWizard interface. Most of the MSBuild XML
generation, which is common to all types of project, will now be located
in this base class, and can then be customized as needed in overrides
of the specific wizard classes.

Task-number: QTVSADDINBUG-589
Task-number: QTVSADDINBUG-737
Change-Id: Ib98d02d7358b589f9174aa74e02c9fdd0dcee668
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>