Fix QHash visualizer not working with const keys
The debug visualizer for QHash was modified to allow const key types to
be used. Previously, when using const key types, the visualizer was
unable to correctly interpret the data structure(*) and would default to
raw view.

By default, the new implementation of the visualizer will show a simple
list of key-value entries. Alternatively, a detailed view of bucket
contents can be shown by specifying the "view(buckets)" option appended
to the variable name in the watch window.

(*) due to an apparent bug in the VS debug visualization, which adds a
trailing space to the key type if it is specified as const, and will not
recognize references to the key type that do not contain the trailing
space.

Task-number: QTVSADDINBUG-643
Change-Id: I952249fba5b1467e0b1202e80c47d5c03709b013
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>