Allow custom path for Qt properties file
Qt settings are introduced in the build process in the form of MSBuild
properties with values extracted from a qmake-generated Makefile. The
file containing the Qt properties is, by default, stored in a sub-folder
of $(IntDir). This default sub-folder, along with the name of the
property file can now be overridden by the user, e.g. by adding to the
project a property sheet containing the customized values.

Task-number: QTVSADDINBUG-721
Change-Id: Iba69f889439bc4bd7889f8745872624fd8995026
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>