Refactor/rewrite QMake class
The QMake class, which encapsulates the logic of executing qmake when
needed (e.g. for generating VS projects or querying qmake properties),
has been refactored/rewritten. The goal of this change was to have more
readable and extensible code, and also to remove the use of concurrent
threads to handle data from the stdout and stderr streams of the qmake
process. This multi-threaded approach resulted in qmake error messages
being shown out-of-sync with the rest of the process output.

Task-number: QTVSADDINBUG-648
Change-Id: I29fa0cdb9d140c365ba07852cfa285762d35c0ec
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>