Keep IntelliSense up to date when opening Qt projects
Occasionally, after opening or creating a project, IntelliSense will fail
to highlight Qt types in C++ code, indicating that Qt build settings,
fetched from a qmake-generated Makefile, were missing at the time of the
design-time build. The generation of the Makefile is now triggered in
the first event received from the VC project system (the ProjectAdded
event for the Solution is triggered before the Opened event).

Change-Id: I68feca1f9ec9b3919a70b90519114bb89669e405
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>