Hide conversion functions (part 2)
Completion of a6b70b670d32141ccf68973ac86d29fa7fc74b78:

"Hide conversion functions between qmake created project and vs tools
project

The conversion broke in recent refactorings. As fixing these will not
be trivial and we do not want to postpone the release for this issue,
the functionality is removed for now and will be brought back later."

Must also hide options in the project context menu.

Change-Id: If11082977b9ce7a78deb36c0f7ee3776b0d19c92
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>