Fix .cpp files with Q_OBJECT macro not handled correctly
When adding a .cpp file containing the Q_OBJECT macro, the default
value of "Dynamic C++ Source" in the moc property page is kept
unchanged as "output". This causes the .cpp file to be incorrectly
excluded from the build, and is inconsistent with the behavior when
importing qmake projects.

Task-number: QTVSADDINBUG-792
Change-Id: I42a8dd27ca7e73f8c568e7fa4c4b1d7bc6b452d3
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>