Register Linux-based Qt installations
Using the Qt Options dialog to register a Qt installation located in a
Linux build host is now possible. The registered version can then be
selected in the project creation wizards or in the Qt build settings
property page.

When registering a Linux-based Qt installation, the access mode -- SSH
or WSL -- must be specified, as well as the path to the installation
directory. Optionally, the path to the C++ compiler can also be
specified (default is g++).

Task-number: QTVSADDINBUG-739
Change-Id: Ib610ddd8d8be135201ea1b13cf59943c2567d62a
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>