Remove WSL path fixup
The MSBuild VC Linux targets for WSL assumes that paths are rooted in
the /mnt directory. This leads to errors when using host-based absolute
paths in build properties. However, fixing these on the Qt/MSBuild
targets is not the correct approach; paths should be assumed to be
correct as provided. The user should be advised to workaroung the issue
of "/mnt" prefixing by MSBuild, e.g. by creating symlinks in /mnt to
other root directories, as required.

Change-Id: I31f2ba604d0ca3e228b3b6e2e901cce06d42afba
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>