Integrate windeployqt with Qt/MSBuild
The Qt for Windows deployment tool (windeployqt) can now be executed at
the end of a project build. This feature is disabled by default. To
enable it a corresponding flag property must be set in the Qt build
settings property page for the project. A dedicated property page for
windeployqt options will then become available. During the build, after
linking is complete, windeployqt will be called with the options
specified in its property page.

Task-number: QTVSADDINBUG-578
Change-Id: I922ad0e20e899d7ea0c3da2ce65e918e47a98f25
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>