Build from command line
Created a .bat script to build the extension from the command line.
The script will build the QtVsTools solution in every installed version
of VS. This is useful, especially, when preparing a new release.

This script can also be used to run an initial text template processing
(with the -init option) in a clean solution. This is useful to avoid
project load errors in VS due to missing generated files after cloning
the VsTools repository.

Task-number: QTVSADDINBUG-735
Change-Id: Id12a4adbd44548ffa3943b1ee2ddb2b1a90578b1
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>