Rework handling of project filters
Filters in VC projects are intended as a means to organize the project
files according to their type (i.e. extension). It is the users'
prerogative to create or remove filters in their projects; extensions
should not try to enforce the existence of specific filters with
non-standard semantics. As a consequence of the revised approach to
project filters, this change fixes errors occurring while adding Qt
resource files to a project.

Also in this change: fixed widget project template adding .ui file to
the Resource Files filter instead of the Form Files filter.

Task-number: QTVSADDINBUG-810
Task-number: QTVSADDINBUG-835
Change-Id: Ia305ba4ac7686abdff7162a97de19c22cc279dce
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>