Fix wait dialog in IntelliSense refresh
Fixed a problem in the refresh of IntelliSense information where, in
VS2017, the request to display a wait dialog and status bar message
would fail and cause the IntelliSense refresh to be skipped completely.
This problem was linked to the way in which the VS service interfaces
for the wait dialog and the status bar were obtained.

Change-Id: Iea44dbc04ab08dd7072218ec8c5498c31519a3cd
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>