Configure QML debug connection timeout
Added the option to configure the timeout for debugger connections to
the QML runtime. The user can set the timeout in msecs or disable the
timeout, in which case the debugger will wait indefinitely for a
connection to the QML runtime.

This is useful when a blocking operation is carried out before
starting the QML runtime, e.g. a modal message box is shown at the
beginning of main(), before calling the application's exec().

Task-number: QTVSADDINBUG-844
Change-Id: I7c7535d1cee9b8b687ebec96ae41badaf0ddb2e2
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>