Reference QtMsBuild.csproj in main project
Files in the Qt/MSBuild project will now be packaged into the VSIX
without requiring explicit entries in the package manifest. For this
purpose, a reference to QtMsBuild was added to the QtVsTools project.
This causes files in QtMsBuild marked as "CopyToOutput" to be added
automatically to the extension package.

This change allows Qt/MSBuild files to be referenced only once in the
QtMsBuild project. Previously, these files were referenced in several
locations throughout the Qt VS Tools solution.

Change-Id: Ic1419740a241c373f11e643cac9837af261695c1
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>