Compile Qt/MSBuild inline tasks during build
Qt/MSBuild inline tasks are now located in separate C# source files,
part of the QtMsBuild project. The C# source is then copied to
qt_tasks.targets when building the add-in. In this way, inline
tasks are compiled before packaging, which allows compilation errors to
be detected at build time.

Compiling inline code into a QtMsBuild assembly also allows the
possibility of running unit tests on the inline tasks.

Change-Id: I38a2ea2b2f4a820f22858ab8ff19f2b4f00a7e37
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>