Query Qt docs on F1 pressed
When the VS online help keyboard shortcut (default: F1) is pressed,
the documentation for the current Qt version is checked. If a relevant
page is found it is displayed. Otherwise, the default online help
behavior of VS is resumed. This 'hook' on the F1 key-press event can be
disabled in the Qt Options page.

Also, added a read-only field to the Qt Options page, to show the
current keyboard shortcut for the Qt online help (default: Alt+F1).

Change-Id: I28e00a7972b9bde8ecc0c8cb8830e6f0311964a8
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>