Refactor WaitDialog
Added default parameter values to the public methods of WaitDialog, to
make calling code more readable. Also, this class will now inherit from
IDisposable, making it possible to use a WaitDialog inside the context
of a `using` block, and having the Close method be called automatically
when leaving that context. This allows writing code like the following:

  using (WaitDialog.Start(...)) {
      // VS wait dialog is now shown
      // Do some UI-blocking processing
      ...
  }
  // VS wait dialog is now closed

Change-Id: Ia8850b7bfad69ce5c0a8f05c137e9a90e00ef5b2
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>