Fix .ts files ignored by Qt translation tools
Fixed a bug where .ts files with item type "None" would not be picked up
by the Qt translation tools. The issue was introduced in the recent
reworking of the integration of translation tools with Qt/MSBuild. This
caused lupdate and lrelease to stop running in existing projects.

Task-number: QTVSADDINBUG-888
Change-Id: Iacbc800f5ef2bbbd20cd4beb164691ec3fdaf59d
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>