Reset macros between test sessions
The C# macros loaded in QtVSTest during a test session will be removed
after that session is over. Previously, macros would remain loaded,
which caused errors and unexpected behavior.

A test session is defined as starting when a test client connects to
the QtVSTest local socket, and ending when that client disconnects.

Change-Id: I965d3a6c68e85e77cb9e61ea680ca005c148dc8a
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>