Allow testing a running instance of VS
Modified the C# macro client interface to allow connecting to and
testing a running instance of VS. This will allow the same instance of
VS to be used across test sessions. Previously, it was only possible to
run a test session on a VS instance created specifically for that
single session.

Change-Id: I8f7a8fd1274911c23b2f70783b3b9a34cab80249
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>