Add test client header to QtVsTest package
The auto-test/C# macro client interface header file, MacroClient.h,
is now included in the QtVsTest extension package. The header file is now
copied to %LOCALAPPDATA% when starting VS after installing QtVsTest.

This change will allow development of auto-tests for VS without requiring
that the QtVSTools repository be cloned. The only requirement will be the
installation of the QtVsTest extension.

Change-Id: Ibbcb52c9068e793ea68f4c40b61903891c4c0a9c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>