Use specific VSSDK assembly location in T4
For VS 2019 and 2017, references in T4 text templates to the VSSDK
assembly Microsoft.VisualStudio.Shell.Interop will now use a path to
the DLL instead of relying on the GAC to locate the assembly by name.
VS 2015 adds the assembly to the GAC, but 2019 and 2017 do not.

This fixes an issue in 4cbf4a4537a11e489da68130a5f608c6664211f3 where
the vstools.bat build script would fail due to T4 text transformation
not being able to locate the VSSDK assembly.

Change-Id: Ia1636747990c7e783b411aa8ebb70aecf7b22d43
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>