Use more precise project event source
In the Qt/MSBuild project tracker, project change events are now sourced
from the ProjectSubscription service. This allows more precision in the
selection of events to receive. Previously, the generic ProjectChanged
event was used, which is raised "after any write lock is released that
caused a change to this project". This was causing the project tracker
to trigger unnecessary background builds.

Task-number: QTVSADDINBUG-890
Change-Id: If27c975b474d9dc2adc1a4335964341506727b9d
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>