Add PriorityQueue type
Added a new generic utility type, PriorityQueue, that represents a
thread-safe collection of objects that can be retrieved by the order of
their priority at enqueue. The underlying representation is based on a
hash-table, mapping items to their priority, and a binary search tree
containing the items sorted by priority. This results in the following
time complexity profile:
 * Enqueue, Dequeue ≈ O(log n)
 * Peek, Contains ≈ O(1)

Also added a specialized PunisherQueue, based on PriorityQueue, which
behaves like the standard Queue<T> of the .NET framework, with the
following main differences:
 * Thread-safe: can be simultaneously used by several threads.
 * Re-queue: Adding the same object (i.e. an object already present in
 the queue) many times will not create multiple entries in the queue for
 that object. Instead, at each call to Enqueue(), the object is moved
 from its current position to the back of the queue (i.e. "punished").

Change-Id: I3e53db1d64190f0470edf645f20312a0e8976ba4
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>