Add Test solution configuration
Added a new 'Test' build configuration to the Qt VS Tools solution.
This can be selected to build and run test projects with release
settings. As such, test projects will now no longer be built as part of
the 'Release' solution configuration.

This change also fixes an issue where the vstools.bat build script
would fail when running with the -init flag. This issue was introduced
as part of the following changes:
  * 15779cabfdc324bef56e04b2e943a81410e4e3e4
  * db326690ec282f06757cacce40710281122e31d8

Change-Id: I7dd2fbb12fe93c83f9c95521eb6cdf698ed7596d
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>