Build sample test project outside of VS Tools solution
The sample test project, previously named 'CoreFeature', and which is
renamed to 'SampleTest' in this change, can now be built outside of the
context of the Qt VS Tools solution. For this purpose, a dedicated
VS solution and qmake .pro file are provided to open the sample test
as a standalone project, respectively, in Visual Studio and Qt Creator.

Change-Id: Icbfa8d19fbc0fc0cf15e889bd78fb2c5116716db
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>