Add concurrency test
Added a sample solution that makes it easier to reproduce concurrency
issues in parallel builds of Qt/MSBuild projects. A .bat script builds
the sample solution in a loop and records any errors. This script can be
left to run unattended in order to find concurrency issues that occur
infrequently.

Change-Id: I006f8436289caf7e8d784fe73c2d1cb5c2ec73e0
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>