Fix possible thread starvation issue
When a target in Qt/MSBuild requests access to a critical section,
a timeout of 10 seconds will now be enforced to prevent a starvation
scenario, in case of leak of the critical section (e.g. the target that
owns the critical section was terminated without running the fail-safe
release of the critical section).

Fixes: QTVSADDINBUG-917
Change-Id: Ie357152ac6ec208d33a5acfa07a45ace0b1f041f
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>