Fix tracking of unloaded projects, null access on tracker object
The action of unloading a project (right click the project, Unload)
did indeed set the project tracker to null, though the project name
as key remained in the concurrent dictionary. Reloading the project
then invokes Get and we pick the now null tracker out of the collection.
Instead of just setting the tracker to null, remove the whole pair and
let it re-add on project reload.

Change-Id: Ia668ee8abef3323eb1128ce214714de3ee8e83a6
Reviewed-by: Miguel Costa <miguel.costa@qt.io>