Fix missing SQLite DLL's
The move from packages.config to PackageReference's removed the SQLite
native interop DLL's from the VSIX. This caused the online help for Qt
types to stop working.

In the previous NuGet integration, via a packages.config file, packages
were installed locally at the solution dir. With the move to
PackageReference's, these are now installed inside the NuGet global
package folder. As such, the SQLite files must be sourced from the
corresponding sub-directory within that root folder.

The condition "'$(PkgStub_System_Data_SQLite_Core_NetFramework)' != ''"
checks if VS has defined this property containing the path to the SQLite
package installation folder. This is only valid for VS2022 and VS2019.

For VS2017, the SQLite folder is calculated with the assumption that the
global package folder is located at $(USERPROFILE)\.nuget\packages. It
is possible to override this default by defining a 'NuGetGlobalPkgDir'
environment variable with the actual path. When support for VS2017 is
eventually dropped, this calculation can then be removed.

Change-Id: I3d9c35b911a1d256f423bef1fee4ba5b031df337
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>