Extend error message printed to output window
In case an editor cannot start because of a missing file, print
the file name to output window too. Current behavior is to only
print the exception message and stack trace, missing the actual
file. Might help with QTVSADDINBUG-868 since the message box
shown by Visual Studio can be misleading, it will always show the
file the user was to open.

Task-number: QTVSADDINBUG-868
Change-Id: Icc7e980a125054edd57723b4f80a7187c60944d2
Reviewed-by: Miguel Costa <miguel.costa@qt.io>