Check if Qt translation tools are available
Qt/MSBuild targets that start lupdate or lrelease will check if the
translation tools are present in the Qt binaries folder. If they're not,
a warning is generated. This will not stop the build from progressing.

Also, menu options related to Qt translation tools are now disabled if
the tools are missing for the active configuration.

Fixes: QTVSADDINBUG-975
Change-Id: Ibe1b993ea0de4acc87a2c59027eab34d47ee8223
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>