Add new Qt/MSBuild task - QtRunTask
Added to Qt/MSBuild a new task, QtRunTask, which wraps the call to
another task, specified by an assembly path and type name. It's assumed
that the specified task accepts a list of source items. Optionally, the
specified task can also generate an output based on the source items.
This output is added as new metadata to the source items. The list of
modified items is returned as the final result.

Task-number: QTVSADDINBUG-979
Change-Id: I3e0abef9f26353208e7abd50ea6f0417a8296660
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>