Fix error when Qt path is invalid
Fixed a bug in the Qt versions options page where, if the path to a Qt
installation was incorrect (i.e. not containing qmake.exe or
bin\qmake.exe), applying changes (pressing 'OK') was allowed but the
Qt version would not be registered or would be removed. This occurred
silently; the only feedback was an exception notification in the output
window. However, this was a generic system exception (Argument exception
in System.IO.Path.Combine()) which did not provide any information about
the cause of the error.

Pressing 'OK' on the versions page will now trigger a validation of the
Qt installation path, and will only proceed if the path is valid.

Fixes: QTVSADDINBUG-973
Change-Id: Ide6187bab1a0616d785d8348cbbd436ff771d8e1
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>