Get version from repo tag
The active version of the Qt VS Tools is now determined exclusively by
the repository tag that precedes the current HEAD. Previously, this
information was written both in the repo tags and also a file in the
solution (versions.targets). With this change, there is now a single
source of truth regarding the version information.

Also in this change: the deployment of release packages (copying of the
.vsix file to a deployment dir) is now done in the vstools.bat script.
Previously, this was done in a post-build target of the Package project.

Change-Id: If4d1b285689cfc3684e76339b8cd1a92703ea9ed
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>
(cherry picked from commit 14c4ae513c6ae7bc6d2c71e7b9319a953017d621)
Reviewed-by: Miguel Costa <miguel.costa@qt.io>