Fix Qt Modules settings page adding a trailing ';'
If the Qt Modules line edit was empty, String.Split did still
return this empty value and it was added later on in the code
by String.Join, leading to a trailing ';'.

Fix QtMsBuild QtModules comparison by cleaning up the list of
modules first, basically like QtBkup_QtModules is created.

Fixes: QTVSADDINBUG-1092
Change-Id: Ia51c712bdf845ccf43c9a40117c4d334679f927e
Reviewed-by: Miguel Costa <miguel.costa@qt.io>