Ensure thread-safety of LazyFactory
Reworked the LazyFactory class to ensure its effective thread-safety.

The thread-safety of LazyFactory relied on the ConcurrentDictionary and
Lazy .NET classes. However, according to the MS documentation, the
GetOrAdd method that was previously used "is not atomic with regards to
all other operations on the ConcurrentDictionary<TKey,TValue> class".
There is the potential risk that the lazy-initialization function
initFunc is called more than once for the same object being added. This
defeats the purpose of using the Lazy<object> class to ensure the
thread-safety of object initialization.

Change-Id: Ie9cf710ed0afd42b409d56dd547ff011f425e218
Reviewed-by: Miguel Costa <miguel.costa@qt.io>