Improve readability of qmake log output
Made changes to the output of qmake and related processes that is
printed to the output window, in order to make the information more
readable. In particular, when running multiple instances of qmake in
parallel processes (e.g. at startup, while loading information about Qt
version), output messages will now consistenly identify the process ID
that generated the original output.

Change-Id: Ib3b182740192653dac32110d065d3b2d52e8e365
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>