qt6.natvis: big improvement to QHash and QMultiHash
Displays the keys/values from a QHash in the debugger.
Unfortunately it's not possible to iterate the buckets and the values of
buckets (we are limited to one iteration), so the solution is to test
the whole 127 possible values in a bucket, the reason for all those
repeated lines.

Task-number: QTVSADDINBUG-991
Change-Id: I7aa3fd3b76a305047d68d531d3515cd58f230ecd
Reviewed-by: Miguel Costa <miguel.costa@qt.io>