Remove support for projects with custom build steps
The QtVersion class now supports only V3 of the project format.
This enables us to remove support for projects with custom build
steps. It is also not necessary anymore to check the project
format before usinq QtProject, it is either a valid object and
we know it is V3 and above, or it failed to create.

Change-Id: I1cb0d8efe06cd8e2afe2fad8b9178c40c379ea7a
Reviewed-by: Miguel Costa <miguel.costa@qt.io>