cmake: Enable QML debug in user preset
QML debug is now enabled in Qt/CMake projects through user presets
instead of the lists file. This fixes the issue of disabled QML debug
for projects not created by vstools (e.g. example projects). Also, it is
arguably a user-specific setting, in which case it should reside in a
user presets file.

Fixes: QTVSADDINBUG-1106
Change-Id: I89351425550ef291cb8944288f2f972985afbded
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>