Check if debug session started during init
If a debug session was started while the extended package initialization
was on-going, will now attach the QML debugger if needed. Previously, if
a debug session was started before init was complete (for example, by
starting VS, opening a project, and pressing F5 while the extension is
still checking registered Qt versions), the debugged process would block
on the connection to the QML debugger, which would never happen.

Change-Id: Ia90399d594e853b906775ec55524fd860f8da2cc
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>