Add utility functions
These functions will be needed for the generation of project format
conversion reports.

 * ToZipBase64: encode a string in UTF8, then compress (Deflate) the
resulting buffer and return it as a Base64-encoded string.

 * FromZipBase64: decode a string generated with ToZipBase64 into the
corresponding original content.

 * LastIndexOfArray: find the last occurrence of an array contained
within another.

 * Hash: calculate the SHA256 of a given buffer, possibly augmented
with an additional piece of text. This can be used as a crude way of
'salting' the hash in order to secure it from possible replay exploits.

Change-Id: I851746e980c575a97757a53df4ee4d1177744b3d
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>