Generate project format conversion report
Generated conversion reports show which files were modified during the
conversion, and allow viewing differences of pre- and post-conversion
file contents. Reports will also show changes made during each step
(i.e. call to Commit()) of the conversion procedure.

Reports are formatted as XAML files, containing a FlowDocument object
that can be displayed on a WPF rich-text control. Additional metadata is
added to allow viewing pre-conversion file contents (which makes the
conversion report also an effective backup of the project).

When a project is converted (including from an imported .pro file), a
conversion report file is generated next to the project file and with
the same name, but with the extension '.qtvscr'. The user is notified of
the successful conversion, and given the option to open the report. For
that purpose, an editor for '.qtvscr' files (ConversionReportViewer) is
also provided. Generated reports can thus be viewed at any time by
opening them in the VS IDE (provided that the Qt VS Tools extension
remains installed).

Change-Id: Id5f360b85b51feddb020c131e4b013aea97f9bb0
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>