Raise exception if legacy extension is detected
During package initialization, if a legacy vstools extension has been
installed and is detected, an exception is then raised. This ensures
that the current and legacy extensions will not be running side-by-side.

This "future-proof" change imposes the following conditions on an
eventual implementation of a legacy support version of the extension:
 - Package GUID for VS 2022 must be 6E7FA583-5FAA-4EC9-9E90-4A0AE5FD61EE
 - Package GUID for VS 2019 must be C3A906C1-8557-4730-8DFD-33966FAD84F5
 - Name of AsyncPackage-derived class must be QtVsToolsLegacyPackage

Change-Id: I4cb4200d604e5c05f2794c928de9bf921f690f0e
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>