Connect to QML runtime only in debug mode
When starting a QML project in a debug configuration, but without debug
enabled (e.g. Ctrl+F5), the parameters for connecting to the QML runtime
should not be passed to the process being started. Otherwise, the
process will block indefinitely as the QML runtime waits for a
connection from the (disabled) debug engine.

Instead of passing the connection parameters as arguments to the QML
process being started, they will instead be passed in an environment
variable (QML_DEBUG_ARGS). A "debug launch hook" will then detect that
debug mode is enabled and will add the connection parameters to the list
of process arguments (i.e. argv). If starting the process without debug
enabled, the launch hook is not invoked and the connection parameters
are correctly omitted from argv, in which case the QML runtime will not
block waiting for a connection.

Using a debug launch hook seems to be the only mechanism available in VS
to detect that debug mode is enabled *before* the debuggee process is
started. All other notifications mechanisms are triggered after the
process was already started, which is too late to make changes to its
argument list.

Fixes: QTVSADDINBUG-1135
Change-Id: I912ca47832bf5fb511e7901a26e7d7f3c02512ea
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>