Fix Qt modules not propely selected after non-Qt project conversion
When converting a non-Qt project, the properties are not explicitly
added to the QtSettings property group. Instead, they rely on the
default properties set by 'qt_default.props'. This can result in an
issue when modifying a property later from the Qt modules property
page UI, causing the default module not to be selected.

Change-Id: Iaabe8912307267cf94c6b4faff787b19713a51b7
Reviewed-by: Miguel Costa <miguel.costa@qt.io>