Fix button state logic on the wizard config page
* Always store the initial state of the 'Next' and 'Finish' button.
* Do not overwrite the initial state of the buttons once we have
  configured a new Qt version (if there was non before).

When starting certain wizards, such as the {Empty, Console, QML} wizard,
there is no subsequent page after the configuration page. The logic for
the 'Next' and 'Finish' button was broken, especially when no Qt version
was configured. Attempting to set up a Qt version using the error
message (a label one can click on) resulted in unnecessary overwriting
of the initial state of the 'Next' and 'Finish' button in the
ErrorMsg_OnMouseDown() function.

Change-Id: I2efbe631bcb567997e349418016c0174cbb4a076
Reviewed-by: Miguel Costa <miguel.costa@qt.io>
Reviewed-by: Ali Can Demiralp <ali.demiralp@qt.io>