The .obj files of generated sources are placed next to their sources
The objects of generated files, such as MOC files were being placed in
the root of the build directory. This was causing problems for example
when a project included source files with the same name in different
directories. This is no longer the case. Object files of generated
source files are now placed in the same directory as their sources.

Change-Id: I063f32c5628999ac20cb57cdee5ecb8f3fa632b1
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>
Reviewed-by: Miguel Costa <miguel.costa@qt.io>