Fix package InitializeAsync call delay
Specified some additional auto-load contexts for the AsyncPackage, to
handle the case where VS is started with the option to immediately load
a solution/project.

Also, the wait-event that signals initialization complete is now static
and async. An await for initialization was added to the QML LSP client,
to avoid the consequences of race conditions between MEF and package
initialization.

Change-Id: Ib065e12be1ace5d6275bf178d5ce2e1fc504b7ec
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>