Ensure pre-compiled header is accessible from generated sources
The project directory was removed from the additional include
directories passed to generated sources in
3746c90a3ee718342ce7f1f4415a07635c7cd1fc due to raising an error about
inconsistency of arguments passed to different CL tasks. However it is
necessary for the generated sources within the build directory to be
able to locate the pre-compiled header in the source directory. This
commit adds the project directory to all CL tasks, preventing the
warning as well as allowing access to the precompiled header.

Change-Id: Ia1cddbf89c258d0c2e43b90f355ea29a7bf4508d
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>
Reviewed-by: Miguel Costa <miguel.costa@qt.io>