Avoid modifying .vcxproj.user
Removed QtLastBackgroundBuild, which was a property that was written,
but never read, and had the sole purpose of triggering an update to
MOC/UIC/... as well as IntelliSense. Replaced it with QtTouchProperty,
which serves the same purpose, but without modifying the .vcxproj.user.
Added comments clarifying its purpose to avoid unintentional refactoring
in the future.

Fixes: QTVSADDINBUG-935
Change-Id: I4d72c9930f07b121fe6437965e7c082b335cf34a
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>