Make use of the new settings class
This patch addresses an initialization issue that occurs when
the user waits on the splash screen until Visual Studio is fully
initialized. In this scenario, none of the options on the
QtOptionsPage are set to their corresponding values from the
registry, which may result in incorrect behavior in some cases.

Change-Id: I9403289f429244ee4f2f6656f703cdb1f60e486a
Reviewed-by: Miguel Costa <miguel.costa@qt.io>