Correct conversion of moc_predef steps and set output path to $(IntDir)
The project importer was ignoring moc_predef steps since it was
searching for "moc.exe" within their commands, which does not exist,
since predefs are handled by cl and not moc.exe. This is no longer the
case. Furthermore the predef output was hard-coded to "debug\" and
"release\" directories under the project root. This has been changed to
respect the $(IntDir).

Fixes: QTVSADDINBUG-987
Change-Id: Ibc6b01d68a7cd27bad3c3148a0562ef399ac743b
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>