Squish: Stabilize building CMake based projects
To determine whether MSVS is ready for building a project, the
test checks the text of a label. Before this change, the execution
proceeded when the label did not display a generic placeholder
text. This was fragile because for an instant, the label might
still display the previous project's name.

Now, the test will wait for the current project's actual name.

Change-Id: I6b9fa93050be49f7dfd4f7778563a83c420ff948
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>