Remove architecture dependent dependency to Microsoft.Build.Utilities.Core.dll
This and other dlls are no longer available inside the architecture
dependent folder, but only in the generic folder one level above. This
leads to build errors when building Qt projects on Windows on ARM.

Depending on $(MSBuildToolsPath) brings a hard dependency to the used
architecture because this path contains the architecture.

According to
https://developercommunity.visualstudio.com/t/ARM64:-MicrosoftBuildUtilitiesCoredl/10574258
the "generic" location for these dependencies should be used.
As Microsoft.Build.Utilities.Core is added to the build environment
automatically we can get rid of this dependency completely.

Fixes: QTVSADDINBUG-1199
Change-Id: Ie28fd52f185f24475bab5c1073586462e094e28f
Reviewed-by: Ali Can Demiralp <ali.demiralp@qt.io>
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>