Squish: Update tests for verifying VS Tools' (un)installation
MSVS2022 17.10.0 comes with a new extension manager which opens in
an editor tab instead of a dialog. It also has a different layout.

This only fixes the verification tests. The semiautomatic tests
which actually (un)install the VS Tools still need to be updated.

The tests still try using the old extension manager which makes
them compatible with older versions, including MSVS2019. If that
fails, they try using the new extension manager.

Change-Id: I219e94b9ea38ab84a2a310740f33f7ecc9280a21
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>